correlation surface: move to tqec.computation.correlation
